+++ Commands/Graph/GetDeletedMicrosoft365Group.cs
@@ -9,24 +9,24 @@ using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 
 namespace SharePointPnP.PowerShell.Commands.Graph
 {
-    [Cmdlet(VerbsCommon.Get, "PnPDeletedUnifiedGroup")]
-
-    [CmdletHelp("Gets one deleted Microsoft 365 Group (aka Unified Group) or a list of deleted Microsoft 365 Groups",
+    [Cmdlet(VerbsCommon.Get, "PnPDeletedMicrosoft365Group")]
+    [Alias("Get-PnPDeletedUnifiedGroup")]
+    [CmdletHelp("Gets one deleted Microsoft 365 Group or a list of deleted Microsoft 365 Groups",
         Category = CmdletHelpCategory.Graph,
         SupportedPlatform = CmdletSupportedPlatform.Online)]
     [CmdletExample(
-        Code = "PS:> Get-PnPDeletedUnifiedGroup",
+        Code = "PS:> Get-PnPDeletedMicrosoft365Group",
         Remarks = "Retrieves all deleted Microsoft 365 Groups",
         SortOrder = 1)]
     [CmdletExample(
-        Code = "PS:> Get-PnPDeletedUnifiedGroup -Identity 38b32e13-e900-4d95-b860-fb52bc07ca7f",
+        Code = "PS:> Get-PnPDeletedMicrosoft365Group -Identity 38b32e13-e900-4d95-b860-fb52bc07ca7f",
         Remarks = "Retrieves a specific deleted Microsoft 365 Group based on its ID",
         SortOrder = 2)]
     [CmdletMicrosoftGraphApiPermission(MicrosoftGraphApiPermission.Group_ReadWrite_All | MicrosoftGraphApiPermission.Group_Read_All)]
-    public class GetDeletedUnifiedGroup : PnPGraphCmdlet
+    public class GetDeletedMicrosoft365Group : PnPGraphCmdlet
     {
         [Parameter(Mandatory = false, HelpMessage = "The Identity of the Microsoft 365 Group")]
-        public UnifiedGroupPipeBind Identity;
+        public Microsoft365GroupPipeBind Identity;
 
         protected override void ExecuteCmdlet()
         {