+++ Commands/Graph/RemoveDeletedMicrosoft365Group.cs
@@ -7,26 +7,25 @@ using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 
 namespace SharePointPnP.PowerShell.Commands.Graph
 {
-    [Cmdlet(VerbsData.Restore, "PnPDeletedUnifiedGroup")]
-
-    [CmdletHelp("Restores one deleted Microsoft 365 Group (aka Unified Group)",
+    [Cmdlet(VerbsCommon.Remove, "PnPDeletedMicrosoft365Group")]
+    [Alias("Remove-PnPDeletedUnifiedGroup")]
+    [CmdletHelp("Permanently removes one deleted Microsoft 365 Group",
         Category = CmdletHelpCategory.Graph,
-        OutputTypeLink = "https://docs.microsoft.com/graph/api/directory-deleteditems-restore",
         SupportedPlatform = CmdletSupportedPlatform.Online)]
     [CmdletExample(
-        Code = "PS:> Restore-PnPDeletedUnifiedGroup -Identity 38b32e13-e900-4d95-b860-fb52bc07ca7f",
-        Remarks = "Restores a deleted Microsoft 365 Group based on its ID",
+        Code = "PS:> Remove-PnPDeletedMicrosoft365Group -Identity 38b32e13-e900-4d95-b860-fb52bc07ca7f",
+        Remarks = "Permanently removes a deleted Microsoft 365 Group based on its ID",
         SortOrder = 1)]
     [CmdletExample(
-        Code = @"PS:> $group = Get-PnPDeletedUnifiedGroup -Identity 38b32e13-e900-4d95-b860-fb52bc07ca7f
-PS:> Restore-PnPDeletedUnifiedGroup -Identity $group",
-        Remarks = "Restores the provided deleted Microsoft 365 Group",
+        Code = @"PS:> $group = Get-PnPDeletedMicrosoft365Group -Identity 38b32e13-e900-4d95-b860-fb52bc07ca7f
+PS:> Remove-PnPDeletedMicrosoft365Group -Identity $group",
+        Remarks = "Permanently removes the provided deleted Microsoft 365 Group",
         SortOrder = 2)]
     [CmdletMicrosoftGraphApiPermission(MicrosoftGraphApiPermission.Group_ReadWrite_All)]
-    public class RestoreDeletedUnifiedGroup : PnPGraphCmdlet
+    public class RemoveDeletedMicrosoft365Group : PnPGraphCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true, HelpMessage = "The Identity of the deleted Microsoft 365 Group")]
-        public UnifiedGroupPipeBind Identity;
+        public Microsoft365GroupPipeBind Identity;
 
         protected override void ExecuteCmdlet()
         {
@@ -34,7 +33,7 @@ PS:> Restore-PnPDeletedUnifiedGroup -Identity $group",
 
             if (group != null)
             {
-                UnifiedGroupsUtility.RestoreDeletedUnifiedGroup(group.GroupId, AccessToken);
+                UnifiedGroupsUtility.PermanentlyDeleteUnifiedGroup(group.GroupId, AccessToken);
             }
         }
     }